Added FINAL gps sensor code
This final gps sensor code does NOT have a terminating ';' I could not figure out how to add it I tried basically everything and for some reason it did not work. Other than that though this code gets a GPS packet (that is currently terminated by a comma ',') about every 4 or 5 seconds. -Landon and David